+++ core-navigation/sample/src/main/java/ru/surfstudio/android/core/navigation/sample/ui/screen/main/MainPresenter.kt
@@ -5,6 +5,8 @@ import ru.surfstudio.android.core.mvp.presenter.BasePresenterDependency
 import ru.surfstudio.android.core.navigation.sample.R
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
 import ru.surfstudio.android.core.navigation.sample.ui.screen.profile.ProfileRoute
+import ru.surfstudio.android.core.ui.navigation.customtabs.CustomTabsNavigator
+import ru.surfstudio.android.core.ui.navigation.customtabs.OpenUrlRoute
 import ru.surfstudio.android.dagger.scope.PerScreen
 import ru.surfstudio.android.core.ui.provider.resource.ResourceProvider
 import javax.inject.Inject
@@ -16,6 +18,7 @@ import javax.inject.Inject
 internal class MainPresenter @Inject constructor(
         basePresenterDependency: BasePresenterDependency,
         private val activityNavigator: ActivityNavigator,
+        private val customTabsNavigator: CustomTabsNavigator,
         private val resourceProvider: ResourceProvider
 ) : BasePresenter<MainActivityView>(basePresenterDependency) {
 
@@ -28,5 +31,9 @@ internal class MainPresenter @Inject constructor(
 
     fun openProfileScreen() {
         activityNavigator.start(ProfileRoute(resourceProvider.getString(R.string.user_name)))
+    }
+
+    fun openLink(link: String) {
+        customTabsNavigator.openLink(OpenUrlRoute(link))
     }
 }