+++ mvp/sample-mvp-dialog/src/main/java/ru/surfstudio/android/mvp/dialog/sample/ui/screen/dialogs/complex/ComplexDialogFragment.kt
@@ -28,8 +28,7 @@ class ComplexDialogFragment : CoreDialogFragmentView() {
         return inflater.inflate(R.layout.complex_dialog_layout, container)
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?, viewRecreated: Boolean) {
         increase_btn.setOnClickListener { presenter.increment() }
         decrease_btn.setOnClickListener { presenter.decrement() }
         apply_btn.setOnClickListener { presenter.applyChanges() }