+++ permission/lib-permission/src/main/java/ru/surfstudio/android/core/ui/permission/PermissionManager.kt
@@ -15,65 +15,58 @@
  */
 package ru.surfstudio.android.core.ui.permission
 
+import android.app.Activity
 import android.content.SharedPreferences
-import android.os.Build
-import android.util.Log
-import androidx.annotation.RequiresApi
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
-
-import java.util.HashMap
-
-import ru.surfstudio.android.core.ui.event.ScreenEventDelegateManager
-import ru.surfstudio.android.core.ui.event.result.RequestPermissionsResultDelegate
-import ru.surfstudio.android.core.ui.provider.ActivityProvider
-
 import androidx.core.content.PermissionChecker.PERMISSION_GRANTED
 import io.reactivex.Completable
+import io.reactivex.Observable
 import io.reactivex.Single
-import io.reactivex.SingleEmitter
-import io.reactivex.SingleTransformer
-import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
-import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityWithResultRoute
+import io.reactivex.disposables.Disposables
+import ru.surfstudio.android.activity.holder.ActiveActivityHolder
 import ru.surfstudio.android.core.ui.permission.exceptions.PermissionsRationalIsNotProvidedException
 import ru.surfstudio.android.core.ui.permission.exceptions.SettingsRationalIsNotProvidedException
 import ru.surfstudio.android.core.ui.permission.screens.default_permission_rational.DefaultPermissionRationalRoute
 import ru.surfstudio.android.core.ui.permission.screens.settings_rational.DefaultSettingsRationalRoute
+import ru.surfstudio.android.navigation.command.activity.Start
+import ru.surfstudio.android.navigation.observer.ScreenResultObserver
+import ru.surfstudio.android.navigation.observer.command.activity.RequestPermission
+import ru.surfstudio.android.navigation.observer.command.activity.StartForResult
+import ru.surfstudio.android.navigation.observer.executor.AppCommandExecutorWithResult
+import ru.surfstudio.android.navigation.observer.route.ActivityWithResultRoute
+import ru.surfstudio.android.navigation.observer.route.PermissionRequestRoute
+import ru.surfstudio.android.navigation.observer.route.ResultRoute
+import ru.surfstudio.android.navigation.route.activity.ActivityRoute
+import ru.surfstudio.android.navigation.rx.extension.observeScreenResult
 import java.io.Serializable
 
 /**
  * Класс для проверки и запросов разрешений.
  */
-abstract class PermissionManager(
-        eventDelegateManager: ScreenEventDelegateManager,
-        private val activityProvider: ActivityProvider,
-        private val activityNavigator: ActivityNavigator,
-        private val sharedPreferences: SharedPreferences
-) : RequestPermissionsResultDelegate {
-
-    private val singleEmitterPerRequestCode = HashMap<Int, SingleEmitter<Boolean>>()
+open class PermissionManager(
+    private val activeActivityHolder: ActiveActivityHolder,
+    private val commandExecutor: AppCommandExecutorWithResult,
+    private val sharedPreferences: SharedPreferences,
+    private val screenResultObserver: ScreenResultObserver
+) {
+    private var settingsDisposable = Disposables.disposed()
+    private var rationalDisposable = Disposables.disposed()
+    private var permissionDisposable = Disposables.disposed()
 
     init {
-        eventDelegateManager.registerDelegate(this)
+        //что бы очистить старые результаты
+        subscribeToSettingsResult()
+        subscribeToRationalResult()
+        subscribeToPermissionResult()
     }
 
-    /**
-     * Способ выдачи разрешений
-     */
-    protected abstract fun performPermissionRequest(permissionRequest: PermissionRequest)
-
-    override fun onRequestPermissionsResult(
-            requestCode: Int,
-            permissions: Array<String>,
-            grantResults: IntArray
-    ): Boolean {
-        val nonNullSingleEmitter = singleEmitterPerRequestCode[requestCode] ?: return false
-        if (nonNullSingleEmitter.isDisposed) return true
-
-        val isPermissionRequestGranted = isAllResultsAreGranted(grantResults)
-        nonNullSingleEmitter.onSuccess(isPermissionRequestGranted)
-
-        return true
+    protected fun performPermissionRequest(permissionRequest: PermissionRequest): Single<Boolean> {
+        val route = PermissionRequestRoute(permissionRequest.permissions)
+        return screenResultObserver
+            .observeScreenResult(route)
+            .firstOrError()
+            .doOnSubscribe { commandExecutor.execute(RequestPermission(route)) }
     }
 
     /**
@@ -83,11 +76,8 @@ abstract class PermissionManager(
      *
      * @return Статус разрешения [PermissionStatus].
      */
-    fun check(permissionRequest: PermissionRequest): PermissionStatus = when {
-        isPermissionRequestGranted(permissionRequest) -> PermissionStatus.GRANTED
-        !isPermissionRequestRequested(permissionRequest) -> PermissionStatus.NOT_REQUESTED
-        isPermissionRequestDenied(permissionRequest) -> PermissionStatus.DENIED
-        else -> PermissionStatus.DENIED_FOREVER_OR_ONE_TIME_PERMISSION
+    fun check(permissionRequest: PermissionRequest): PermissionStatus {
+        return checkInternal(permissionRequest).blockingGet()
     }
 
     /**
@@ -98,215 +88,191 @@ abstract class PermissionManager(
      * @return [Single], содержащий [Boolean]: true, если разрешение выдано, false - если нет.
      */
     fun request(permissionRequest: PermissionRequest): Single<Boolean> {
-        val permissionRequestStatus = check(permissionRequest)
-
-        if (permissionRequestStatus.isGranted) {
-            return Single.just(true)
-        }
-
-        return showPermissionRequestRationalIfNeeded(permissionRequest)
-                .toSingleDefault(false)
-                .flatMap {
-                    performPermissionRequestByDialogOrSettings(permissionRequest, permissionRequestStatus)
+        disposeAll()
+        return checkInternal(permissionRequest)
+            .flatMap { permissionRequestStatus ->
+                if (permissionRequestStatus.isGranted) {
+                    Single.just(true)
+                } else {
+                    showPermissionRequestRationalIfNeeded(permissionRequest)
+                        .toSingleDefault(false)
+                        .flatMap {
+                            performPermissionRequestByDialogOrSettings(
+                                permissionRequest,
+                                permissionRequestStatus
+                            )
+                        }
+                        .doOnSuccess {
+                            setPermissionRequestIsRequested(permissionRequest)
+                            setPermissionRequestIsGranted(permissionRequest, it)
+                        }
                 }
-                .doOnSuccess {
-                    setPermissionRequestIsRequested(permissionRequest)
-                }
+            }
     }
 
-    private fun isAllResultsAreGranted(grantResults: IntArray): Boolean =
-            grantResults
-                    .all { grantResult -> grantResult == PERMISSION_GRANTED }
+    private fun checkInternal(permissionRequest: PermissionRequest): Single<PermissionStatus> {
+        return isPermissionRequestGranted(permissionRequest).flatMap { isGranted ->
+            val status = when {
+                isGranted -> PermissionStatus.GRANTED
+                !isPermissionRequestRequested(permissionRequest) -> PermissionStatus.NOT_REQUESTED
+                isPermissionRequestLastGranted(permissionRequest) -> PermissionStatus.GRANTED_ONCE
+                isPermissionRequestDenied(permissionRequest) -> PermissionStatus.DENIED
+                else -> PermissionStatus.DENIED_FOREVER_OR_ONE_TIME_PERMISSION
+            }
+            Single.just(status)
+        }
+    }
 
-    /**
-     * Проверяем, выдал ли пользователь разрешения.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
-    private fun isPermissionRequestGranted(permissionRequest: PermissionRequest): Boolean =
-            permissionRequest
-                    .permissions
-                    .all { permission -> isPermissionGranted(permission) }
+    private fun subscribeToSettingsResult() {
+        val route = DefaultSettingsRationalRoute("")
+        settingsDisposable = screenResultObserver
+                .observeScreenResult(route)
+                .firstElement()
+                .subscribe()
+    }
 
-    /**
-     * Проверяем, отказал ли пользователь в выдаче разрешений.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
+    private fun subscribeToRationalResult() {
+        val route = DefaultPermissionRationalRoute("")
+        rationalDisposable = screenResultObserver
+                .observeScreenResult(route)
+                .firstElement()
+                .subscribe()
+    }
+
+    private fun subscribeToPermissionResult() {
+        val route = PermissionRequestRoute(emptyArray())
+        permissionDisposable = screenResultObserver
+            .observeScreenResult(route)
+            .firstOrError()
+            .subscribe({}, {})
+    }
+
+    private fun disposeAll() {
+        settingsDisposable.dispose()
+        permissionDisposable.dispose()
+        rationalDisposable.dispose()
+    }
+
+    private fun isPermissionRequestGranted(permissionRequest: PermissionRequest): Single<Boolean> =
+        Observable.fromArray(*permissionRequest.permissions)
+            .flatMapSingle { isPermissionGranted(it) }
+            .all { it }
+
     private fun isPermissionRequestDenied(permissionRequest: PermissionRequest): Boolean =
-            shouldShowRequestPermissionRationale(permissionRequest)
+        shouldShowRequestPermissionRationale(permissionRequest)
 
-    /**
-     * Проверяем запрашивались ли уже разрешения.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
-    private fun isPermissionRequestRequested(permissionRequest: PermissionRequest): Boolean  = permissionRequest
+    private fun isPermissionRequestRequested(permissionRequest: PermissionRequest): Boolean =
+        permissionRequest
             .permissions
             .all { permission -> isPermissionRequested(permission) }
 
-    /**
-     * Отображает диалог, объясняющий, зачем требуется это разрешение, если нужно.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
     private fun showPermissionRequestRationalIfNeeded(permissionRequest: PermissionRequest): Completable =
-            if (needToShowPermissionRequestRational(permissionRequest)) {
-                showPermissionRequestRational(permissionRequest)
-            } else {
-                Completable.complete()
-            }
-
-    /**
-     * Если разрешение было выдано, то пользователь получит true, иначе, для версий Android до 11,
-     * мы предлагаем пользователю перейти в настройки, на версии после 11, мы вызываем
-     * tryOneTimePermissionRequest.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     * @param permissionStatus статус [PermissionRequest].
-     */
-    private fun performPermissionRequestByDialogOrSettings(
-            permissionRequest: PermissionRequest,
-            permissionStatus: PermissionStatus
-    ): Single<Boolean> =
-            when {
-                permissionStatus != PermissionStatus.DENIED_FOREVER_OR_ONE_TIME_PERMISSION ->
-                    performPermissionRequestByDialog(permissionRequest)
-                Build.VERSION.SDK_INT >= Build.VERSION_CODES.R -> tryOneTimePermissionRequest(permissionRequest)
-                permissionRequest.showSettingsRational -> performPermissionRequestBySettings(permissionRequest)
-                else -> Single.just(false)
-            }
-
-    /**
-     * Данный метод нужен для Android 11, из-за одноразовых разрешений флаги в случае, когда пользоатель
-     * выбрал одноразовые разрешения или "Don't ask again" совпадают, поэтому на Android 11 мы вначале пытаемся запросить у системы
-     * разрешение, если разрешение уже запрашивалось и пользователь отказал, то shouldShowRequestPermissionRationale
-     * вернет true, это будет означать, что первый запрос уже был, пользователь отказал и мы пропускаем
-     * дальше по цепочке, если shouldShowRequestPermissionRationale вернет false, то это означает, что
-     * пользователь выбрал "Don't ask again" и надо предложить перейти ему в настройки.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
-    @RequiresApi(Build.VERSION_CODES.R)
-    private fun tryOneTimePermissionRequest(
-            permissionRequest: PermissionRequest
-    ): Single<Boolean> = performPermissionRequestByDialog(permissionRequest).flatMap { isGranted ->
-        //Данное условие определит, если в настройках указанно, что у нас только одноразовое разрешение
-        if (isGranted || (!isGranted && shouldShowRequestPermissionRationale(permissionRequest))) {
-            Single.just(isGranted)
+        if (needToShowPermissionRequestRational(permissionRequest)) {
+            showPermissionRequestRational(permissionRequest)
         } else {
-            if (permissionRequest.showSettingsRational) {
-                performPermissionRequestBySettings(permissionRequest)
-            } else {
-                Single.just(false)
-            }
+            Completable.complete()
         }
-    }
 
-    /**
-     * Установить флаг того, что у пользователя запрашивали разрешения.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
+    private fun performPermissionRequestByDialogOrSettings(
+        permissionRequest: PermissionRequest,
+        permissionStatus: PermissionStatus
+    ): Single<Boolean> =
+        when {
+            permissionStatus != PermissionStatus.DENIED_FOREVER_OR_ONE_TIME_PERMISSION ->
+                performPermissionRequest(permissionRequest)
+            permissionRequest.showSettingsRational -> performPermissionRequestBySettings(
+                permissionRequest
+            )
+            else -> Single.just(false)
+        }
+
     private fun setPermissionRequestIsRequested(permissionRequest: PermissionRequest) =
-            permissionRequest
-                    .permissions
-                    .forEach { permission -> setPermissionIsRequested(permission) }
+        permissionRequest
+            .permissions
+            .forEach { permission -> setPermissionIsRequested(permission) }
 
-    /**
-     * Проверить, давал ли пользователь разрешение.
-     *
-     * @return true, если пользователь дал разрешение, false, если пользователь не выдавал разрешение
-     */
-    private fun isPermissionGranted(permission: String): Boolean =
-            ContextCompat.checkSelfPermission(activityProvider.get(), permission) == PERMISSION_GRANTED
+    private fun setPermissionRequestIsGranted(
+        permissionRequest: PermissionRequest,
+        isGranted: Boolean
+    ) = permissionRequest
+        .permissions
+        .forEach { permission -> setPermissionIsGranted(permission, isGranted) }
 
-    /**
-     * Проверить, следует ли показывать пользователю объяснение, для чего нужны запрашиваемые Permissions.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     *
-     * @return True, если во время предыдущего запроса разрешения пользователь нажал отказ и не выбрал опцию "Don't ask
-     * again", false - если разрешение запрашивается в первый раз, или если во время предыдущего запроса разрешения
-     * пользователь выбрал опцию "Don't ask again".
-     */
-    private fun shouldShowRequestPermissionRationale(permissionRequest: PermissionRequest): Boolean {
-        return permissionRequest
-                .permissions
-                .any { permission -> shouldShowPermissionRationale(permission) }
-    }
+    private fun isPermissionGranted(permission: String): Single<Boolean> =
+        safeGetActivity().map {
+            ContextCompat.checkSelfPermission(
+                it,
+                permission
+            ) == PERMISSION_GRANTED
+        }
 
-    /**
-     * Проверяет запрашивали ли мы ранее это разрешение.
-     *
-     * @param permission Проверяемое разрешение.
-     *
-     * @return true если уже запрос по этому разрешению был, false если не было запроса.
-     */
-    private fun isPermissionRequested(permission: String) = sharedPreferences.getBoolean(permission, false)
+    private fun shouldShowRequestPermissionRationale(permissionRequest: PermissionRequest): Boolean =
+        permissionRequest
+            .permissions
+            .any { permission -> shouldShowPermissionRationale(permission).blockingGet() }
+
+    private fun isPermissionRequested(permission: String) =
+        sharedPreferences.getBoolean(permission, false)
+
+    private fun isPermissionRequestLastGranted(permissionRequest: PermissionRequest): Boolean =
+        permissionRequest
+            .permissions
+            .all { permission -> isPermissionLastGranted(permission) }
+
+    private fun isPermissionLastGranted(permission: String) =
+        sharedPreferences.getBoolean(GRANTED_PREFIX + permission, false)
+
+    private fun setPermissionIsGranted(permission: String, isGranted: Boolean) =
+        sharedPreferences
+            .edit()
+            .putBoolean(GRANTED_PREFIX + permission, isGranted)
+            .apply()
 
     private fun needToShowPermissionRequestRational(permissionRequest: PermissionRequest): Boolean =
-            permissionRequest.showPermissionsRational && shouldShowRequestPermissionRationale(permissionRequest)
+        permissionRequest.showPermissionsRational && shouldShowRequestPermissionRationale(
+            permissionRequest
+        )
 
-    /**
-     * Отображает диалог, объясняющий, зачем требуется это разрешение.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
     private fun showPermissionRequestRational(permissionRequest: PermissionRequest): Completable {
         val customPermissionsRationalRoute = permissionRequest.permissionsRationalRoute
         val customPermissionsRationalStr = permissionRequest.permissionsRationalStr
 
         val permissionRationalRoute = when {
-            customPermissionsRationalRoute != null -> customPermissionsRationalRoute
-            customPermissionsRationalStr != null -> DefaultPermissionRationalRoute(customPermissionsRationalStr)
+            customPermissionsRationalRoute != null -> {
+                return startAndObserveReturnFromScreenCustom(customPermissionsRationalRoute)
+            }
+            customPermissionsRationalStr != null -> DefaultPermissionRationalRoute(
+                customPermissionsRationalStr
+            )
             else -> return Completable.error(PermissionsRationalIsNotProvidedException())
         }
 
         return startAndObserveReturnFromScreen(permissionRationalRoute)
     }
 
-    /**
-     * Отображает системный диалог запроса разрешения
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
-    private fun performPermissionRequestByDialog(permissionRequest: PermissionRequest): Single<Boolean> =
-            Single
-                    .create<Boolean> { singleEmitter ->
-                        singleEmitterPerRequestCode[permissionRequest.requestCode] = singleEmitter
-                        performPermissionRequest(permissionRequest)
-                    }
-                    .doFinally { singleEmitterPerRequestCode.remove(permissionRequest.requestCode) }
-
-    /**
-     * Предлагаем пользователю перейти на экран настроек приложения для активации разрешения.
-     *
-     * @param permissionRequest Проверяемый [PermissionRequest].
-     */
     private fun performPermissionRequestBySettings(permissionRequest: PermissionRequest): Single<Boolean> {
         val customSettingsRationalRoute = permissionRequest.settingsRationalRoute
         val customSettingsRationalStr = permissionRequest.settingsRationalStr
 
         val settingsRationalRoute = when {
-            customSettingsRationalRoute != null -> customSettingsRationalRoute
-            customSettingsRationalStr != null -> DefaultSettingsRationalRoute(customSettingsRationalStr)
+            customSettingsRationalRoute != null -> {
+                return startAndObserveReturnFromScreenCustom(customSettingsRationalRoute)
+                    .andThen(checkInternal(permissionRequest).map { it.isGranted })
+            }
+            customSettingsRationalStr != null -> DefaultSettingsRationalRoute(
+                customSettingsRationalStr
+            )
             else -> return Single.error(SettingsRationalIsNotProvidedException())
         }
         return startAndObserveReturnFromScreen(settingsRationalRoute)
-                .toSingle { check(permissionRequest).isGranted }
+            .andThen(checkInternal(permissionRequest).map { it.isGranted })
     }
 
-    /**
-     * Устанавливает флаг, что проверяемое разрешение запрошенно.
-     *
-     * @param permission Проверяемое разрешение.
-     */
     private fun setPermissionIsRequested(permission: String) =
-            sharedPreferences
-                    .edit()
-                    .putBoolean(permission, true)
-                    .apply()
+        sharedPreferences
+            .edit()
+            .putBoolean(permission, true)
+            .apply()
 
     /**
      * Проверить, следует ли показывать пользователю объяснение, для чего нужен запрашиваемый Permission.
@@ -317,13 +283,37 @@ abstract class PermissionManager(
      * again", false - если разрешение запрашивается в первый раз, или если во время предыдущего запроса разрешения
      * пользователь выбрал опцию "Don't ask again".
      */
-    private fun shouldShowPermissionRationale(permission: String) =
-            ActivityCompat.shouldShowRequestPermissionRationale(activityProvider.get(), permission)
+    private fun shouldShowPermissionRationale(permission: String): Single<Boolean> =
+        safeGetActivity().map {
+            ActivityCompat.shouldShowRequestPermissionRationale(it, permission)
+        }
 
-    private fun startAndObserveReturnFromScreen(route: ActivityWithResultRoute<*>): Completable =
-            activityNavigator
-                    .observeResult<Serializable>(route)
-                    .firstElement()
-                    .flatMapCompletable { Completable.complete() }
-                    .doOnSubscribe { activityNavigator.startForResult(route) }
+    private fun startAndObserveReturnFromScreenCustom(route: ActivityRoute): Completable {
+        return screenResultObserver
+            .observeScreenResult(route as ResultRoute<Serializable>)
+            .firstElement()
+            .flatMapCompletable { Completable.complete() }
+            .doOnSubscribe { commandExecutor.execute(Start(route)) }
+    }
+
+    private fun <T : Serializable> startAndObserveReturnFromScreen(route: ActivityWithResultRoute<T>): Completable =
+        screenResultObserver
+            .observeScreenResult(route)
+            .firstElement()
+            .flatMapCompletable { Completable.complete() }
+            .doOnSubscribe { commandExecutor.execute(StartForResult(route)) }
+
+    private fun safeGetActivity(): Single<Activity> {
+        val activity = activeActivityHolder.activity
+        return if (activity != null) {
+            Single.just(activity)
+        } else {
+            activeActivityHolder.activityObservable
+                .firstOrError()
+        }
+    }
+
+    private companion object {
+        const val GRANTED_PREFIX = "granted"
+    }
 }