+++ template/run_danger.sh
@@ -1,0 +1,2 @@
+bundle install
+bundle exec danger